+++ CHANGELOG.md
@@ -59,6 +59,12 @@
 [10353]: https://github.com/enso-org/enso/pull/10353
 [10396]: https://github.com/enso-org/enso/pull/10396
 
+#### Enso Language & Runtime
+
+- Support for [explicit --jvm option][10374] when launching `enso` CLI
+
+[10374]: https://github.com/enso-org/enso/pull/10374
+
 #### Enso Standard Library
 
 - [Added Statistic.Product][10122]